move component ui creation into re_component_ui
