+++ crates/viewer/re_data_ui/src/component_ui_registry.rs
@@ -1,29 +1,21 @@
-use re_chunk_store::{LatestAtQuery, RowId};
-use re_entity_db::EntityDb;
-use re_log_types::{external::arrow2, EntityPath};
-use re_types::external::arrow2::array::Utf8Array;
 use re_ui::UiExt;
-use re_viewer_context::{ComponentUiRegistry, UiLayout, ViewerContext};
+use re_viewer_context::ComponentUiRegistry;
 
 use super::EntityDataUi;
 
-pub fn create_component_ui_registry() -> ComponentUiRegistry {
+pub fn register_component_uis(registry: &mut re_viewer_context::ComponentUiRegistry) {
     re_tracing::profile_function!();
 
-    let mut registry = ComponentUiRegistry::new(Box::new(&fallback_component_ui));
+    // TODO(#6661): Move this to component_ui_registry. Separate components could simplify this to the extent that multi/single line is enough?
+    add_to_registry::<re_types::components::AnnotationContext>(registry);
 
-    // TODO(andreas): Move this to component_ui_registry. Separate components could simplify this to the extent that multi/single line is enough?
-    add_to_registry::<re_types::components::AnnotationContext>(&mut registry);
+    // TODO(#6661): Move this to component_ui_registry. Image preview is a bit hard because row_id and size stuff needs to be known
+    add_to_registry::<re_types::components::Blob>(registry);
+    add_to_registry::<re_types::components::TensorData>(registry);
 
-    // TODO(andreas): Move this to component_ui_registry. Image preview is a bit hard because row_id and size stuff needs to be known
-    add_to_registry::<re_types::components::Blob>(&mut registry);
-    add_to_registry::<re_types::components::TensorData>(&mut registry);
-
-    // TODO(andreas): Move this to component_ui_registry. Funky AnnotationContext querying thing. Maybe we can get away with a store querying hack?
-    add_to_registry::<re_types::components::ClassId>(&mut registry);
-    add_to_registry::<re_types::components::KeypointId>(&mut registry);
-
-    registry
+    // TODO(#6661): Move this to component_ui_registry. Funky AnnotationContext querying thing. Maybe we can get away with a store querying hack?
+    add_to_registry::<re_types::components::ClassId>(registry);
+    add_to_registry::<re_types::components::KeypointId>(registry);
 }
 
 /// Registers how to show a given component in the ui.
@@ -60,64 +52,4 @@ pub fn add_to_registry<C: EntityDataUi + re_types::Component>(registry: &mut Com
             },
         ),
     );
-}
-
-#[allow(clippy::too_many_arguments)]
-fn fallback_component_ui(
-    _ctx: &ViewerContext<'_>,
-    ui: &mut egui::Ui,
-    ui_layout: UiLayout,
-    _query: &LatestAtQuery,
-    _db: &EntityDb,
-    _entity_path: &EntityPath,
-    _row_id: Option<RowId>,
-    component: &dyn arrow2::array::Array,
-) {
-    arrow_ui(ui, ui_layout, component);
-}
-
-fn arrow_ui(ui: &mut egui::Ui, ui_layout: UiLayout, array: &dyn arrow2::array::Array) {
-    use re_types::SizeBytes as _;
-
-    // Special-treat text.
-    // Note: we match on the raw data here, so this works for any component containing text.
-    if let Some(utf8) = array.as_any().downcast_ref::<Utf8Array<i32>>() {
-        if utf8.len() == 1 {
-            let string = utf8.value(0);
-            ui_layout.data_label(ui, string);
-            return;
-        }
-    }
-    if let Some(utf8) = array.as_any().downcast_ref::<Utf8Array<i64>>() {
-        if utf8.len() == 1 {
-            let string = utf8.value(0);
-            ui_layout.data_label(ui, string);
-            return;
-        }
-    }
-
-    let num_bytes = array.total_size_bytes();
-    if num_bytes < 3000 {
-        // Print small items:
-        let mut string = String::new();
-        let display = arrow2::array::get_display(array, "null");
-        if display(&mut string, 0).is_ok() {
-            ui_layout.data_label(ui, &string);
-            return;
-        }
-    }
-
-    // Fallback:
-    let bytes = re_format::format_bytes(num_bytes as _);
-
-    // TODO(emilk): pretty-print data type
-    let data_type_formatted = format!("{:?}", array.data_type());
-
-    if data_type_formatted.len() < 20 {
-        // e.g. "4.2 KiB of Float32"
-        ui_layout.data_label(ui, &format!("{bytes} of {data_type_formatted}"));
-    } else {
-        // Huge datatype, probably a union horror show
-        ui.label(format!("{bytes} of data"));
-    }
 }
